1.fix some problem 2.add douyu/huya/egame live play 3.version 1.0.0.14 Signed-off-by: y673893131 <673893132@qq.com>

+++ video_player_core/video_thread.h
@@ -7,7 +7,7 @@
 #include <unistd.h>
 #define msleep(x) usleep(x * 1000)
 #elif(WIN32)
-#define msleep(x) Sleep(x)
+#define msleep(x) Sleep(static_cast<unsigned long>(x))
 #endif
 
 class video_thread
@@ -15,7 +15,7 @@ class video_thread
 public:
     virtual ~video_thread();
     static void start(const _video_info_&);
-    static video_thread* index(int);
+    static video_thread* index(size_t);
     void play();
     void setPause();
     bool getPause();
